pyproject.toml
[project]
name = "yt-playlist-audio-extractor-with-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.7",
    "moviepy>=2.2.1",
    "python-dotenv>=1.1.1",
    "pytube>=15.0.0",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "yt-dlp>=2025.6.30",
]
